+++ docs/release-notes.md
@@ -5,6 +5,9 @@
 * For players:
   * Fixed `log_context` command not disabling the extra logs when run again.
 
+* For external tools:
+  * Added toolkit method to read the compatibility list from a local copy of its Git repo.
+
 ## 4.1.10
 Released 18 December 2024 for Stardew Valley 1.6.14 or later.